10k_analysis/requirements.txt
azure-search-documents>=11.4.0
openai>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pandas>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0